requirements.txt
pandas
numpy
SQLAlchemy
psycopg2-binary
python-dotenv
